requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.116.0
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.6.1
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0

# WebSocket support
websockets==15.0.1

# Database  
psycopg==3.1.18
sqlalchemy>=2.0.25,<2.1.0

# Google ADK and AI
google-adk==1.9.0
google-cloud-aiplatform==1.95.1

# Audio processing
pyaudio==0.2.14
wave==0.0.2
numpy
scipy

# Environment and utilities
python-dotenv==1.0.0
jsonschema==4.20.0
tzdata>=2024.1

# AWS SDK
boto3==1.35.24
